requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
mysql-connector-python==9.5.0

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

